+++ src/models/Trace.ts
@@ -1,4 +1,5 @@
 import { NativeAPM } from '../native/NativeAPM';
+import type * as APM from '../modules/APM';
 
 export default class Trace {
   constructor(
@@ -8,9 +9,12 @@ export default class Trace {
   ) {}
 
   /**
-   * Add an attribute with key and value to the Trace to be sent.
-   * @param key
-   * @param value
+   * Adds an attribute with a specified key and value to the Trace to be sent.
+   *
+   * @param key - The key of the attribute.
+   * @param value - The value of the attribute.
+   *
+   * @deprecated Please migrate to the App Flows APIs: {@link APM.startFlow}, {@link APM.endFlow}, and {@link APM.setFlowAttribute}.
    */
   setAttribute(key: string, value: string) {
     NativeAPM.setExecutionTraceAttribute(this.id, key, value);
@@ -18,8 +22,11 @@ export default class Trace {
   }
 
   /**
-   * End Execution Trace
+   * Ends the execution trace.
+   *
+   * @deprecated Please migrate to the App Flows APIs: {@link APM.startFlow}, {@link APM.endFlow}, and {@link APM.setFlowAttribute}.
    */
+
   end() {
     NativeAPM.endExecutionTrace(this.id);
   }